Update Readme: Hint to rename
When we tried to manually clone the repo into my moodle/mod directory, Moodle did not find it. As a result, the /moodle/admin/index.php did not react to a newly installed plugin. 

After trying to run a script, we stumbled upon this, because it was looking for "path/opencast/foo.php". We had to rename the plugin to "opencast". Highly appreciated, if you could include this note :)